Refactor playblast dialogs to reuse shared tab UI
